requirements.txt
kafka-python
pyspark
numpy
pandas
python-dotenv
s3fs